NLP/chat_demo/requirements.txt
tensorflow-gpu
jieba
gensim
efweb